requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
openai-whisper>=20231117
pyannote.audio>=3.0.1
pydub>=0.25.1
librosa>=0.10.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
huggingface-hub>=0.17.0
ffmpeg-python>=0.2.0
safetensors>=0.4.0